Moved mod files to a its own folder and file
