+++ src/main.rs
@@ -4,27 +4,7 @@ use std::error::Error;
 type Message = String;
 type Result<T, E = Box<dyn Error>> = std::result::Result<T, E>;
 
-mod files {
-    use std::path::{Path, PathBuf};
-    use std::fs;
-    use super::Result;
-
-    pub fn get_markdown_files<P: AsRef<Path>>(path: P) -> Result<Vec<PathBuf>> {
-        let mut paths = vec![];
-    
-        for result in fs::read_dir(path)? {
-            let path = result?.path();
-    
-            if let Some(ext) = path.extension() {
-                if ext == "md" {
-                    paths.push(path);
-                }
-            }
-        }
-    
-        Ok(paths)
-    }
-}
+use onelo::files;
 
 /// The build command handles the process of parsing all markdown files.
 ///
@@ -84,13 +64,3 @@ fn main() {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_get_only_md_files() {
-        let paths = files::get_markdown_files("test/files");
-        assert_eq!(paths.unwrap().len(), 2);
-    }
-}
